move component metadata to separate tab
