+++ assets/src/components/cd/services/service/ServiceComponents.tsx
@@ -24,8 +24,6 @@ import { useComponentKindSelect } from 'components/apps/app/components/Component
 import { ComponentList } from 'components/apps/app/components/ComponentList'
 import { ModalMountTransition } from 'components/utils/ModalMountTransition'
 
-import { isUnstructured } from 'components/component/ComponentInfo'
-
 import {
   getServiceDetailsBreadcrumbs,
   useServiceContext,
@@ -130,7 +128,6 @@ export default function ServiceComponents() {
               ? `${getServiceComponentPath({
                   clusterId,
                   serviceId,
-                  defaultRaw: isUnstructured(c.kind),
                   componentId: c.id,
                 })}`
               : undefined
