+++ assets/src/routes/cdRoutesConsts.tsx
@@ -163,15 +163,13 @@ export function getClusterAddOnDetailsPath({
 
 export function getServiceComponentPath({
   componentId,
-  defaultRaw = false,
   ...props
 }: Parameters<typeof getServiceDetailsPath>[0] & {
   componentId: string | null | undefined
-  defaultRaw?: boolean
 }) {
   return `${getServiceDetailsPath({
     ...props,
-  })}/${SERVICE_COMPONENTS_PATH}/${componentId}${defaultRaw ? '/raw' : ''}`
+  })}/${SERVICE_COMPONENTS_PATH}/${componentId}`
 }
 
 export function getNodeDetailsPath({
